+++ src/components/profilePage.scss
@@ -1,0 +1,54 @@
+@import '../Styles/GlobalStyles.scss';
+
+
+body {
+    background-color: $blueBackground;
+    color: $whiteFontColor;
+    font-family: $PublicSans;
+}
+
+
+.input-field{
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    p{
+        width: 50px;
+        margin-right: 20px;
+    }
+}
+
+.profile-button-image{
+    width: 100px;
+    height: 100px;
+}
+
+#build-your-profile-button-text{
+    margin-top: 0;
+}
+
+.pop-up-profile-section{
+    display: flex;
+    flex-direction: column;
+    background-color: #EE8434;
+    width: 80%;
+    justify-self: center;
+    justify-content: center;
+    font-family: $MartianMono;
+    #profile-astronaut-picture{
+        width: 100px;
+        height: auto;
+        border: black 2px solid;
+    }
+}
+
+.profile-information{
+    line-height: 25px;
+}
+
+.profile-rover-photos{
+    width: 50%;
+    height: auto;
+    justify-content: center;
+    margin: 0 auto;
+}
